refactor(front): move sandbox DTO types out of lib/api into types/api
Ninth batch of the lib/api DTO migration: internal sandbox endpoints. All three
files were partial (DTOs + business logic); response/request body types move to
types/api/sandbox/*, business logic stays.

- egress_policy: moved Get/PutWorkspaceEgressPolicyResponseBody; kept all
  policy read/write/invalidate logic.
- env_vars: moved Get/PostWorkspaceSandboxEnvVarsResponseBody; dropped the now-
  unused WorkspaceSandboxEnvVarType import; kept all constants/validation.
- image/index.ts: moved SandboxKillImagesResponseBody; kept the barrel
  re-exports and image functions.
- 5 consumers updated; 3 front-api routes split (DTO -> types, functions -> lib).
