+++ front/types/api/sandbox/egress_policy.ts
@@ -1,0 +1,9 @@
+import type { EgressPolicy } from "@app/types/sandbox/egress_policy";
+
+export type GetWorkspaceEgressPolicyResponseBody = {
+  policy: EgressPolicy;
+};
+
+export type PutWorkspaceEgressPolicyResponseBody = {
+  policy: EgressPolicy;
+};
