+++ front/types/api/sandbox/image/index.ts
@@ -1,0 +1,3 @@
+export interface SandboxKillImagesResponseBody {
+  images: Array<{ baseImage: string; version: string }>;
+}
